Fix problem caused by online-check cron task: it attemps to rm .offline even though the file does not exist

+++ bin/online-check.sh
@@ -5,7 +5,9 @@
  
 local offline=`dig 8.8.8.8 +time=1 +short google.com A | grep -c "no servers could be reached"`
 if [[ "$offline" == "0" ]]; then
-  rm ~/.offline
+    if [ -f ~/.offline ]; then
+      rm ~/.offline
+    fi
 else
   touch ~/.offline
 fi
